+++ copasi/layout/CLGroup.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/layout/CLGroup.cpp,v $
-//   $Revision: 1.4.2.1 $
+//   $Revision: 1.4.2.2 $
 //   $Name:  $
 //   $Author: shoops $
-//   $Date: 2010/11/03 17:08:17 $
+//   $Date: 2011/01/12 19:01:27 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -84,7 +84,7 @@ CLGroup::CLGroup(const CLGroup& source, CCopasiContainer* pParent):
 {
   this->mKey = CCopasiRootContainer::getKeyFactory()->add("RenderGroup", this);
   // copy the elements
-  unsigned int i, iMax = source.mElements.size();
+  size_t i, iMax = source.mElements.size();
   CCopasiObject* pChild = NULL;
 
   for (i = 0; i < iMax; ++i)
@@ -195,12 +195,12 @@ CLGroup::CLGroup(const Group& source, CCopasiContainer* pParent):
         break;
     }
 
-  unsigned int i, iMax = source.getNumElements();
+  size_t i, iMax = source.getNumElements();
   const Transformation2D* pChild = NULL;
 
   for (i = 0; i < iMax; ++i)
     {
-      pChild = dynamic_cast<const Transformation2D*>(source.getElement(i));
+      pChild = dynamic_cast<const Transformation2D*>(source.getElement((unsigned int) i));
 
       if (dynamic_cast<const Rectangle*>(pChild))
         {
@@ -379,7 +379,7 @@ const std::string& CLGroup::getEndHead() const
 /**
  * Returns the number of elements in the group.
  */
-unsigned int CLGroup::getNumElements() const
+size_t CLGroup::getNumElements() const
 {
   return this->mElements.size();
 }
@@ -404,7 +404,7 @@ CCopasiVector<CCopasiObject>* CLGroup::getListOfElements()
  * Returns element with index n.
  * If there is no such element, NULL is returned.
  */
-CCopasiObject* CLGroup::getElement(unsigned int n)
+CCopasiObject* CLGroup::getElement(size_t n)
 {
   if (n < this->mElements.size())
     {
@@ -420,7 +420,7 @@ CCopasiObject* CLGroup::getElement(unsigned int n)
  * Returns element with index n.
  * If there is no such element, NULL is returned.
  */
-const CCopasiObject* CLGroup::getElement(unsigned int n) const
+const CCopasiObject* CLGroup::getElement(size_t n) const
 {
   if (n < this->mElements.size())
     {
@@ -596,7 +596,7 @@ void CLGroup::importOldCurve(const XMLNode& node)
 {
     const XMLAttributes& curveAttributes=node.getAttributes();
     const XMLNode* child;
-    unsigned int n=0,nMax = node.getNumChildren();
+    size_t n=0,nMax = node.getNumChildren();
     const XMLNode* pOrigAnnotation=NULL;
     const XMLNode* pOrigNotes=NULL;
     while(n<nMax)
@@ -605,7 +605,7 @@ void CLGroup::importOldCurve(const XMLNode& node)
   std::string childName=child->getName();
   if(childName=="listOfCurveSegments")
   {
-     unsigned int i,iMax=child->getNumChildren();
+     size_t i,iMax=child->getNumChildren();
      const XMLNode* child2=NULL;
      CLRenderPoint start;
      CLRenderPoint end;
@@ -638,7 +638,7 @@ void CLGroup::importOldCurve(const XMLNode& node)
       {
         continue;
       }
-      unsigned int j,jMax=child2->getNumChildren();
+      size_t j,jMax=child2->getNumChildren();
       for(j=0;j<jMax;++j)
       {
         const XMLNode* child3=&child2->getChild(j);
@@ -859,7 +859,7 @@ Group* CLGroup::toSBML(unsigned int level, unsigned int version) const
         break;
     }
 
-  unsigned int i, iMax = this->mElements.size();
+  size_t i, iMax = this->mElements.size();
 
   for (i = 0; i < iMax; ++i)
     {
